+++ src/main/kotlin/com/solve/domain/problem/dto/response/ProblemResponse.kt
@@ -14,7 +14,7 @@ data class ProblemResponse(
     val input: String,
     val output: String,
     val memoryLimit: Long,
-    val timeLimit: Long,
+    val timeLimit: Double,
     var correctRate: Double? = null,
     val testCases: List<ProblemTestCaseResponse>,
     val author: ProblemAuthorResponse,
